\docType{data}
\name{tvec}
\alias{tvec}
\title{1router time data from Cao et al. (2000)}
\description{
  Data from 4-node network with star topology collected
  from Bell Labs; used in Cao et al. (2000). The numeric
  vector \code{tvec} contains the time in decimal hours
  since midnight for each observation.
}
\references{
  J. Cao, D. Davis, S. Van Der Viel, and B. Yu.
  Time-varying network tomography: router link data.
  Journal of the American Statistical Association,
  95:1063–75, 2000.
}
\seealso{
  \code{\link{onerouter}}, \code{\link{X}},
  \code{\link{Y}}, \code{\link{A}}
}
\keyword{datasets}

